Check the PDF file

This function checks that the PDF file does not exist to avoid overwriting a plot the user has previously made.

check_pdf(PDF = "four_panels.pdf", OUTDIR = tempdir())

Arguments

PDF

The name of the PDF file. Defaults to four_panels.pdf.

OUTDIR

The default directory where PDF will be saved to.

Value

The path to the PDF file if the file doesn't exist. It appends the pdf file extension if it was absent from PDF.

Details

This is an utility function used by four_panels and plot_coverage.

Author

Leonardo Collado-Torres

Examples

## Choose a random PDF file
PDF <- paste0("test_", stats::runif(1, max = 1e10))

## Initially this works because the output PDF does not exist.
PDF <- check_pdf(PDF)
## It also adds the PDF extension if the user didn't supply it.
PDF
#> [1] "/tmp/RtmpE58b1z/test_2726604086.34692.pdf"

## Create a dummy PDF file
pdf(file = PDF)
plot(1, 1)
dev.off()
#> agg_png 
#>       2 

## Now it doesn't work since the PDF file already exists.
testthat::expect_error(
    check_pdf(basename(PDF)),
    "already exists! Rename or erase it"
)