Example base-pair region coverage data

A list of base-pair region coverage data.frame lists used for exemplifying the package functionality. This is the data extracted for the example region chr20:10286777-10288069:+ by looping through the phenotype tables stored in pd and using the getRegionCoverage function. This can be reproduced using the brainflowprobes_cov function in this package.

Format

A list with the base-pair coverage output from brainflowprobes_cov with the example region used throughout the package documentation ('chr20:10286777-10288069:+').

Sep

base-pair coverage region data.frame list for pd$Sep,

Deg

base-pair coverage region data.frame list for pd$Deg,

Cell

base-pair coverage region data.frame list for pd$Cell,

Sort

base-pair coverage region data.frame list for pd$Sort.

See also

Examples

if (FALSE) {
    ## Takes about 10 minutes to run!
    four_panels_example_cov <- brainflowprobes_cov(
        "chr20:10286777-10288069:+"
    )
}